update Nomad API to v0.10.0-beta1
